requirements.txt
# this is the requirements .txt file

#movies
python-dotenv
requests
Pillow
Matplotlib